refactor: language classic game score modal
